% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_features.R
\name{feature_new_vars}
\alias{feature_new_vars}
\title{Which new endogenous variables each feature introduces}
\usage{
feature_new_vars(features)
}
\arguments{
\item{features}{Character vector of feature names.}
}
\value{
Character vector of new endogenous variable names.
}
\description{
Which new endogenous variables each feature introduces
}
\keyword{internal}
